rename: Manager 도메인 디렉토리 수정
